% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_servo.R
\name{get_servo}
\alias{get_servo}
\title{Get a servo instance.}
\usage{
get_servo(pin, initial_value = 0)
}
\arguments{
\item{pin}{A character or numeric of length one with the GPIO pin that the
servo is connected to.}

\item{initial_value}{A numeric of length one. If 0 (the default), the
device's mid-point will be set initially. Other values between -1 and +1
can be specified as an initial position. If NULL, it will not move.}
}
\value{
A Servo S4 object, this object will be set to work with the
  configured servo.
}
\description{
Returns a configured servo instance.
}
\examples{
\dontrun{
# Get a servo instance connected at GPIO 14.
servo_14 <- get_servo(14)
# Get a servo instance connected at GPIO 18.
servo_18 <- get_servo(18)
}

}
