% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Servo.R
\docType{class}
\name{Servo-class}
\alias{Servo-class}
\title{Servo Class Representation.}
\description{
Servo Class Representation.
}
\section{Slots}{

\describe{
\item{\code{pin}}{The servo's assigned pin.}

\item{\code{servo}}{The internal representatio of the servo.}
}}

