% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_position.R
\name{set_position}
\alias{set_position}
\title{Set the position of a servo instance.}
\usage{
set_position(x, value)
}
\arguments{
\item{x}{A Servo instance.}

\item{value}{A numeric of length one. The new position of the servo as a
value between -1 (the minimum position) and +1 (the maximum position).}
}
\value{
The same Servo instance provided as input.
}
\description{
Sets the position of a given servo instance.
}
\examples{
\dontrun{
# Get a servo instance.
servo <- get_servo(14)
# Move the servo to its maximum position.
set_position(servo, 1)
# Move the servo to its minimum position.
set_position(servo, -1)
# Move the servo to its midium position.
set_position(servo, 0)
}

}
